BUG: Fix merging display table fields containing commas
If a series description (or other field that is merged with mergeConcatenate) contained a comma, then it was concatenated even if it was the same. For example "first, second" was not found in the split fields because they were "first" and "second". Fixed by simply searching for the new value in the unsplit initial value.

There might be issues with this too, for example if if there are series descriptions like "first", "second", and also "first, second", but in that case seeing only "first, second" is acceptable.