ENH: Reduce memory usage of DICOM indexer
Allow defining cached tags that are not stored in the DICOM tag cache by value. This prevents excessive memory usage and file reading for tags that are not appropriate to store in the tag cache (such as pixel data) but it is still possible to query if the tag is present.

Force updating the database after a certain number of files are parsed. This limits the total memory usage, which could grew big when hundreds of thousands of files were parsed in one batch.

Fixed updating existing images (identified by SOPInstanceUID) when they are loaded from a different file.

Do not show all series when the browser is initially shown, as this can take significant amount of time when a large database is loaded.