ENH: Add checkBoxControlsButtonToggleState property to ctkCheckablePushButton
If checkBoxControlsButtonToggleState is enabled then clicking the checkbox also makes the button pushed (for user's convenience) and unpushing the button makes it unchecked, too (to avoid the ambiguous state when the button's checkbox is checked but the button is not pressed down).

The property is disabled for now to preserve the current behavior (for backward compatibility).

Co-authored-by: Kyle Sunderland <sunderlandkyl@gmail.com>