BUG: Fix font style in DICOM browser
With recent Qt versions, such as Qt-5.15, the special font set in the DICOM browser (in ctkDICOMTableView) does not get rendered correctly on Windows.
It is safer to remove this special font setting in CTK. It can be overridden set by the application anyway.

The only other place where font style was set in CTK was in ctkXnatTreeBrowserMainWindow. For consistency, the font setting has been removed from there, too.

Fixes https://github.com/Slicer/Slicer/issues/4951