ENH: add command and history properties to ctkConsole
Previously, the API allowed execution of commands but not
setting the state of the command buffer to be executed when the
user hit enter.

Also the command history always started empty.

This change adds read/write properties to fix this.

Now the command buffer can be set, e.g. by a hot key or other
custom code and it will be editable and the user can add more text,
user the completer, etc.

Also the command history can be controlled by the appication so
that it can save and restore previous sessions as a convenience
to the user.