ENH: Make DICOM database relocatable
Previously full path was stored to DICOM files that were copied into the DICOM database. This made it impossible to move the database into any other location or store in on removable storage (such as a USB stick, which may be mapped with different path on different computers).

Now files that are copied (not linked) to the DICOM database are stored with path relative to the DICOM database folder (where the .sql files are located).

Also, DICOM database location can now be specified with relative path in application settings. This allows the database to be moved along with the application. For example, to implement a DICOM viewer that opens DICOM database in a folder relative to the application's location (e.g., use the application as a viewer on a DICOM CD).