BUG: Fix DICOM import from within the database folder
When data sets were imported from within the DICOM database folder without copying to the database, the full absolute path was stored in the database.
However, when later the filename was used, it was determined to be a file within the database folder and therefore the filename was converted to relative path.

Fixed the issue by always using relative paths if a data set is imported from within the database folder.

This allows passing ownership of DICOM files to the database without making a copy (by moving the files to anywhere within the database folder and import from there without requesting a copy).