ENH: Make gradient opacity transfer function shift independently from opacity and color transfer functions
X axis of scalar opacity and color transfer functions is image intensity, so it makes sense to shift them together along the x axis.

However, x axis of gradient opacity transfer function is the image intensity gradient, therefore it should not be changed when the image is offset along x axis.