ENH: Change DICOM displayed field update in indexing to use custom rules
The function that updates displayed fields after DICOM indexing uses an internal instance of the database, which does not contain the displayed field generator rules that may be registered externally. The update displayed fields step has been moved to the main thread using the database set to the indexer, so that it has access to the custom rules.