ENH: Display thumbnail in DICOM metadata
To allow a way to reduce "time to first image" in the DICOM browser and help testing troubleshooting, an image thumbnail is displayed in the DICOM metadata browser window.
The thumbnail is only retrieved if the collapsible section is open (to save space and improve performance when accessing large images).

Based on ProjectWeek discussion:
https://projectweek.na-mic.org/PW35_2021_Virtual/Projects/mpReview/

Co-authored-by: Andras Lasso <lasso@queensu.ca>