BUG: avoid app freeze on mac
After recent changes to allow native file browsers,
clicking the button would apparently create a dialog
on mac but it was not visible.  Because it was modal
the app would be alive but unresponive.

This change replaces show() and raise() which work
for QDialog with exec() which is apparently needed
for the native dialog to work properly.

https://discourse.slicer.org/t/slicer-is-freezing-on-macos-when-clicking-on-import-dicom-files/18624

Co-authored-by: Andras Lasso <lasso@queensu.ca>