ENH: Make ctkDirectoryButton elided
Directory paths are often very long, which can force very wide GUI layouts if the full path is displayed.
To avoid this, ctkDirectoryButton now enables eliding by default (display beginning and end of the path with ... in the middle),
so that if the user sets size policy to Expanding then the button will not be wide and will display a correctly shortened path.