ENH: Add elide option to ctkPushButton and allow ctkDirectoryButton and ctkPathLineEdit to shrink horizontally
Add elide option to ctkPushButton: This can be used for displaying a label with long text without enforcing a large minimum size. If elideMode is set then it is recommended to set the horizontal size policy to Expanding.

Make ctkDirectoryButton elided: Directory paths are often very long, which can force very wide GUI layouts if the full path is displayed. To avoid this, ctkDirectoryButton now enables eliding by default (display beginning and end of the path with ... in the middle) and horizontal size policy is expandable, so that the button will not be wide and will display a correctly shortened path.

Make ctkPathLineEdit horizontal sizing size policy to prevent long paths forcing a wide widget (similarly to ctkDirectoryButton).