ENH: Improve robustness of Python code running in the console
When a complete string of command is executed (either copy-pasting or running from file) then we no longer split the string to lines and execute line by line because that often leads to errors. For example, parsing of this Python code line by line would fail due to the empty line in the function definition:

```
def test1():
    print("one")

    print("two")
    print("three")

def test2():
    print("ok")
```

Also make the console remember the path of the last run Python file to make it easier to run it again.