BUG: Fix unnecessary warning in ctkVTKHistogram::resetRange()
When the input image of ctkVTKHistogram was empty then resetRange() logged a warning. An empty image is a valid input (the histogram should be empty) and resetRange() operation is valid on the empty histogram, too. Therefore, it is not necessary to log a warning.

Removing the warning reduces noise during normal operation, which helps in detecting real problems.