BUG: Fixed restoring of DICOM metadata viewer window size
After the DICOM metadata viewer window was show maximized and then closed by hitting Escape button: the metadata viewer opened the widget maximized, but the layout inside was much smaller.

Applied parts of [this workaround](https://stackoverflow.com/questions/44005852/qdockwidgetrestoregeometry-not-working-correctly-when-qmainwindow-is-maximized) to ensure that the widget size is restored when the window is maximized.