COMP: Remove obsolete CTK copy of CMakePackageConfigHelpers
This commit is a follow-up of bb46b8cde (Add support for install-able and
relocatable package.) where a copy of CMakePackageConfigHelpers was integrated
to support configuring using CMake 2.8.7.

This was needed because the module CMakePackageConfigHelpers was introduced
in CMake 2.8.8.

Since CMake 3.0 is now required to build or use CTK, ths commit removes
the obsolete logic checking if the CTK copy of CMakePackageConfigHelpers
needed to be included.

The change is consistent with configure output:

[...]
-- Including CMake built-in module CMakePackageConfigHelpers
-- Including CMake built-in module CMakePackageConfigHelpers - ok
[...]