ENH: applying DICOM display updates even if some fail
Previously, if some data imported into the `ctkDICOMDatabase` had
incomplete information then the display field updates would abort.

With this change, a more informative message is generated to support
troubleshooting of any failed updates, but the rest of the process
completes so that any valid display field updates can be shown
to the user.