ENH: update dicom db schema and api for files and urls
Schema now has URL column of size 2048 (which based on research
is suggested as the max size for URLs).  The schema now
allows either field to be NULL and does not require the
Filename column to be unique.

The api now icnludes a urlsForSeries option to get the list
of URLs for a given series instance uid.

The fileValue method now accepts either filePaths or URLs with
filePaths being given priority if both exist.