ENH: Update ctkLogger adding support for log level
Set ctkLogger default level is set based on the build configuration.
If a Debug build, the default log level is set as `ctkErrorLogLevel::LogLevel::Debug`
and it set as `ctkErrorLogLevel::LogLevel::Warning` otherwise.

The mapping of the logger functions is the following:

* `ctkLogger::trace()` -> `qDebug()
* `ctkLogger::debug()` -> `qDebug()`
* `ctkLogger::info()` -> `qInfo()`
* `ctkLogger::warn()` -> `qWarning()`
* `ctkLogger::error()` -> `qCritical()`
* `ctkLogger::fatal()` -> `qCritical()`

For convenience, the order of the function declaration has been updated
to match the ordering of log level (Trace < Fatal).

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>