ENH: Add method to check if double-spinbox value is being edited
New method is added to ctkDoubleSpinBox and ctkSliderWidget to check if a value is currently being set using the spinbox.
This is useful because when there are many sibling widgets then many signals may be emitted and it may be difficult to determine which widget was the source of all changes.

This feature was needed for fixing https://github.com/Slicer/Slicer/issues/7574