BUG: Fix ctkDICOMDatabase::fileValue for non-cached files
Recent 88ff72b9c2b1e57cf4f822ea2e547be5b9a32b98 commit changed ctkDICOMDatabase::fileValue behavior
so that it no longer returns DICOM tag value for files that are not already in the database.
This is an issue because this method has been used universally for any DICOM files
(it returned result faster for files that are cached in the DICOM database but it worked for files not in the database, too).

This commit restores the previous, preferred behavior.