COMP: Update to find_package(Python3)
The FindPythonInterp and FindPythonLibs modules are removed.

These modules have been deprecated since CMake 3.12.

CMake 3.27 and above prefer to not provide the modules. This policy provides
compatibility for projects that have not been ported away from them.

Projects using the FindPythonInterp and/or FindPythonLibs modules should
be updated to use one of their replacements:
    FindPython3
    FindPython2
    FindPython