COMP: Additional add Q_OBJECT instrumentation
A class with Q_OBJECT inside a local scope makes it a local class.
In C++, local classes cannot declare static data members, and since Q_OBJECT generates a staticMetaObject, this is invalid.

Recommended fix: Instead of declaring a local class within a function, define it at a namespace level.