COMP: Fix ctkDICOMSchedulerTest1 test
This changes the test to check the number of running jobs instead of
the total number of jobs. The previous check was incorrect because
jobs are not automatically destroyed when they finish, leading to
a mismatch in the expected number of jobs.

The updated check ensures that the number of running jobs matches
the number of images. This should generally hold true unless an
image retrieval finishes before the check, making the test
non-deterministic in some cases. We can observe this behavior and
adjust if necessary.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>